+++ Sources/StreamChatUI/ChatChannelList/ChatChannelListItemView_Tests.swift
@@ -29,6 +29,7 @@ class ChatChannelListItemView_Tests: XCTestCase {
         view.content = content
         view.content = nil
         view.addSizeConstraints()
+        view.components = .mock
         AssertSnapshot(view)
     }
     
@@ -36,6 +37,7 @@ class ChatChannelListItemView_Tests: XCTestCase {
         let view = ChatChannelListItemView().withoutAutoresizingMaskConstraints
         view.addSizeConstraints()
         view.content = content
+        view.components = .mock
         AssertSnapshot(view)
     }
     
@@ -48,6 +50,7 @@ class ChatChannelListItemView_Tests: XCTestCase {
         
         view.appearance = appearance
         view.addSizeConstraints()
+        view.components = .mock
         
         AssertSnapshot(view)
     }
@@ -93,6 +96,7 @@ class ChatChannelListItemView_Tests: XCTestCase {
         let view = TestView().withoutAutoresizingMaskConstraints
         
         view.addSizeConstraints()
+        view.components = .mock
         
         view.content = content
         AssertSnapshot(view)