+++ Sources/StreamChatUI/ChatMessageList/Attachments/ChatMessageImageGallery+ImagePreview.swift
@@ -29,7 +29,7 @@ extension ChatMessageGalleryView {
         public var didTapOnAttachment: ((ChatMessageImageAttachment) -> Void)?
         public var didTapOnUploadingActionButton: ((ChatMessageImageAttachment) -> Void)?
 
-        private var imageTask: ImageTask? {
+        private var imageTask: Cancellable? {
             didSet { oldValue?.cancel() }
         }
 
@@ -87,15 +87,15 @@ extension ChatMessageGalleryView {
             let attachment = content
 
             loadingIndicator.isVisible = true
-            imageTask = imageView
-                .loadImage(
-                    from: attachment?.payload.imagePreviewURL,
-                    resize: false,
-                    components: components
-                ) { [weak self] _ in
+            imageTask = components.imageLoader.loadImage(
+                into: imageView,
+                url: attachment?.payload.imagePreviewURL,
+                imageCDN: components.imageCDN,
+                completion: { [weak self] _ in
                     self?.loadingIndicator.isVisible = false
                     self?.imageTask = nil
                 }
+            )
 
             uploadingOverlay.content = content?.uploadingState
             uploadingOverlay.isVisible = attachment?.uploadingState != nil