+++ Sources/StreamChatUI/ChatMessageList/ChatMessage/ChatMessageContentView.swift
@@ -444,12 +444,13 @@ open class ChatMessageContentView: _View, ThemeProvider {
         
         // Avatar
         let placeholder = appearance.images.userAvatarPlaceholder1
-        if let imageURL = content?.author.imageURL {
-            authorAvatarView?.imageView.loadImage(
-                from: imageURL,
+        if let imageURL = content?.author.imageURL, let imageView = authorAvatarView?.imageView {
+            components.imageLoader.loadImage(
+                into: imageView,
+                url: imageURL,
+                imageCDN: components.imageCDN,
                 placeholder: placeholder,
-                preferredSize: .avatarThumbnailSize,
-                components: components
+                preferredSize: .avatarThumbnailSize
             )
         } else {
             authorAvatarView?.imageView.image = placeholder
@@ -502,12 +503,16 @@ open class ChatMessageContentView: _View, ThemeProvider {
 
         // The last thread participant is the author of the most recent reply.
         let threadAvatarUrl = content?.threadParticipants.last?.imageURL
-        threadAvatarView?.imageView.loadImage(
-            from: threadAvatarUrl,
-            placeholder: appearance.images.userAvatarPlaceholder4,
-            preferredSize: .avatarThumbnailSize,
-            components: components
-        )
+
+        if let imageView = threadAvatarView?.imageView {
+            components.imageLoader.loadImage(
+                into: imageView,
+                url: threadAvatarUrl,
+                imageCDN: components.imageCDN,
+                placeholder: appearance.images.userAvatarPlaceholder4,
+                preferredSize: .avatarThumbnailSize
+            )
+        }
 
         // Reactions view
         reactionsBubbleView?.tailDirection = content