+++ Sources/StreamChatUI/ChatMessageList/ChatMessage/ChatMessageContentView_Tests.swift
@@ -18,10 +18,12 @@ final class ChatMessageContentView_Tests: XCTestCase {
     private let createdAt = DateFormatter.Stream.iso8601Date(from: "2019-12-12T15:33:46.488935Z")!
 
     func test_appearance() {
+        let components = Components.mock
+        
         // Iterate over messages and layouts.
         for (message, layout) in testMessagesAndLayouts {
             // Create message content view with the provided `message` and `layout`.
-            let view = contentView(message: message, layout: layout)
+            let view = contentView(message: message, layout: layout, components: components)
 
             // Assert message content view is rendered correctly.
             AssertSnapshot(view, variants: [.defaultLight], suffix: layout.description)