+++ Sources/StreamChatUI/CommonViews/AvatarView/ChatUserAvatarView.swift
@@ -23,15 +23,14 @@ open class ChatUserAvatarView: _View, ThemeProvider {
     }
 
     override open func updateContent() {
-        if let url = content?.imageURL {
-            presenceAvatarView.avatarView.imageView.loadImage(
-                from: url,
-                preferredSize: .avatarThumbnailSize,
-                components: components
-            )
-        } else {
-            presenceAvatarView.avatarView.imageView.image = appearance.images.userAvatarPlaceholder1
-        }
+        components.imageLoader.loadImage(
+            into: presenceAvatarView.avatarView.imageView,
+            url: content?.imageURL,
+            imageCDN: components.imageCDN,
+            placeholder: appearance.images.userAvatarPlaceholder1,
+            preferredSize: .avatarThumbnailSize
+        )
+        
         presenceAvatarView.isOnlineIndicatorVisible = content?.isOnline ?? false
     }
 }