+++ Sources/StreamChatUI/CommonViews/Suggestions/ChatMentionSuggestionView/ChatMentionSuggestionView_Tests.swift
@@ -48,6 +48,7 @@ class ChatMentionSuggestionView_Tests: XCTestCase {
 
     func test_defaultAppearance() {
         let view = ChatMentionSuggestionView().withoutAutoresizingMaskConstraints
+        view.components = .mock
         view.widthAnchor.constraint(equalToConstant: Self.defaultCellWidth).isActive = true
 
         view.content = chatUserOnline
@@ -81,7 +82,7 @@ class ChatMentionSuggestionView_Tests: XCTestCase {
             }
         }
 
-        var components = Components()
+        var components = Components.mock
         components.onlineIndicatorView = RectIndicator.self
         components.mentionAvatarView = CustomAvatarView.self
 
@@ -128,6 +129,7 @@ class ChatMentionSuggestionView_Tests: XCTestCase {
         }
 
         let view = TestView().withoutAutoresizingMaskConstraints
+        view.components = .mock
         view.widthAnchor.constraint(equalToConstant: Self.defaultCellWidth).isActive = true
 
         view.content = chatUserOnline