+++ Sources/StreamChatUI/Utils/Cancellable.swift
@@ -1,0 +1,10 @@
+//
+// Copyright © 2021 Stream.io Inc. All rights reserved.
+//
+
+import Foundation
+
+/// Cancellable provides a set of functions that enable cancelling a task
+public protocol Cancellable {
+    func cancel()
+}